% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_internals.R
\name{dot_internals}
\alias{dot_internals}
\title{Prepend internal functions with a dot}
\usage{
dot_internals()
}
\value{
Changed \code{.R} and \code{.Rmd} files with all instances of internal functions
replaced with their new "dotted" versions.
}
\description{
Prepend internal functions with a dot
}
