% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotInternals-package.R
\docType{package}
\name{dotInternals-package}
\alias{dotInternals}
\alias{dotInternals-package}
\title{dotInternals: Prepend Internal Package Function Names With a Dot}
\description{
Utilities to distinguish non-exported package functions by prepending their names with a dot.
}
\author{
\strong{Maintainer}: Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets) [copyright holder]

}
\keyword{internal}
